modified '/post/:post_id/delete' endpoint to check user's identity before proceeding to post deletion

+++ server/routes/posts.js
@@ -88,16 +88,40 @@ router.patch('/post/:post_id/edit', checkPostOwner, (request, response) => {
 });
 
 router.delete('/post/:post_id/delete', checkPostOwner, (request, response) => {
+	const token = request.headers['authorization'];
 	const requested_post_id = request.params.post_id;
+	const username = request.body.username;
+	const password = request.body.password;
+	const user_id = request.body.user_id;
 
 	if (response.locals.post_ownership) {
-		const delete_post_query = "DELETE FROM `user_posts` WHERE `post_id` = " + requested_post_id;
 
-		database.query(delete_post_query, (error, data) => {
-			if (error) return response.json(error);
+		if (token) {
+			try {
+				const decoded_token = jwt.verify(token, jwt_key);
 
-			response.json({success: true, message: "Post was deleted."});
-		});
+				if (decoded_token) {
+					if (decoded_token.id == user_id) {
+						const delete_post_query = "DELETE FROM `user_posts` WHERE `post_id` = " + requested_post_id;
+
+						database.query(delete_post_query, (error, data) => {
+							if (error) return response.json(error);
+
+							response.json({success: true, message: "Post was deleted."});
+						});
+					} else {
+						response.status(401).json({success: false, message: "Passed token does not correspond to the passed user ID."})
+					}
+				}
+			} catch (error) {
+				console.log(error)
+				response.status(401).json({success: false, message: "Passed token is either invalid, modified or expired."})
+			}
+			
+		} else {
+			response.status(401).json({success: false, message: "No authorization header passed."})
+		}
+
 	} else {
 		response.status(403).json({success: false, message: "Client does not own this post."});
 	}
